Fix milestone_user_watchers migration - use unsignedInteger for milestone_id to match milestones table

+++ database/migrations/2026_03_20_204545_create_milestone_user_watchers_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('milestone_user_watchers', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('milestone_id');
+            $table->unsignedInteger('milestone_id');
             $table->unsignedBigInteger('user_id');
             $table->timestamps();
